pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragsearch-bench"
version = "0.1.0"
description = "Unified CLI and single-venv wrapper around the RAGSearch benchmark (Search-o1, GraphSearch, and friends)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "RAGSearch contributors" }]

# Core benchmark deps. Intentionally CPU-only and free of vllm/CUDA wheels:
# they cover the Search-o1 and GraphSearch *client* paths that talk to a
# remote OpenAI-compatible endpoint.
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24",
    "aiohttp>=3.9",
    "requests>=2.31",
    "tqdm>=4.66",
    "numpy>=1.23",
    "tenacity>=8.2",
    "tiktoken>=0.5",
    "nltk>=3.8",
    "transformers>=4.40",
    "sentencepiece>=0.2",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "pydantic>=2.5",
    "torch",
]

[project.optional-dependencies]
# Retriever backends -- pick only the ones you need.
hipporag  = ["hipporag>=2.0.0a4"]
graphrag  = ["graphrag==1.0.1"]
linearrag = [
    "sentence-transformers>=2.2.2",
    "spacy>=3.6.1",
    "scikit-learn>=1.3",
    "python-igraph>=0.11",
]
raptor    = [
    "faiss-cpu",
    "scikit-learn",
    "umap-learn>=0.5",
]
dense     = [
    "faiss-cpu",
    "datasets",
    "pyserini",
]

# Dev tooling
dev = ["pytest>=7", "ruff>=0.4"]

# GPU-only, for reproducing the RL training paths (Search-R1, Graph-R1).
# Not required to run the benchmark.
train = [
    "vllm>=0.6.3",
    "flash-attn",
    "accelerate",
    "ray",
    "wandb",
]

[project.scripts]
ragbench = "ragsearch_bench.cli:main"

[project.urls]
Homepage = "https://github.com/glenilame2101/RAGSearch"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["ragsearch_bench*"]

[tool.ruff]
line-length = 100
target-version = "py310"
